* 계층간 데이터 전달하기 위한 DTO 적용 * DTO <-> Entity간 팩토리 메소드 제공 * AuditorAware를 통한 등록자, 등록일, 수정자, 수정일 자동 등록 * 등록자, 수정자는 스프링 스큐리티 로그인한 사용자로 자동 등록 * REST API URI 규칙 변경 ( 공통 적인 URL는  @RequestMapping의 value 속성으로 뻄 * 예외 추상화

+++ src/main/java/com/springblog/repository/BlogRepository.java
@@ -3,5 +3,5 @@ package com.springblog.repository;
 import com.springblog.domain.Blog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 }
